Rename directories/files to use camelBack casing
